+++ webknossos/tests/dataset/test_dataset.py
@@ -1,5 +1,6 @@
 import itertools
 import json
+import os
 import pickle
 from pathlib import Path
 from typing import Iterator, Optional, Tuple, cast
@@ -16,10 +17,12 @@ from tests.constants import (
     TESTOUTPUT_DIR,
     use_minio,
 )
+from webknossos.client.context import webknossos_context
 from webknossos.dataset import (
     COLOR_CATEGORY,
     SEGMENTATION_CATEGORY,
     Dataset,
+    RemoteDataset,
     View,
 )
 from webknossos.dataset._array import DataFormat
@@ -1038,6 +1041,17 @@ def test_open_dataset_without_num_channels_in_properties() -> None:
         assert data["dataLayers"][0].get("numChannels") == 1
 
     assure_exported_properties(ds)
+
+
+@pytest.mark.use_proxay
+def test_explore_and_add_remote() -> None:
+    with webknossos_context("http://localhost:9000", os.environ.get("WK_TOKEN")):
+        remote_ds = RemoteDataset.explore_and_add_remote(
+            "http://localhost:9000/data/zarr/Organization_X/l4_sample/",
+            "added_remote_ds",
+            "/Organization_X",
+        )
+        assert remote_ds.name == "added_remote_ds"
 
 
 def test_no_largest_segment_id() -> None: